Add log out and token authentication skeleton
