+++ client/src/index.js
@@ -1,15 +1,22 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { CookiesProvider } from "react-cookie";
+import { createStore, combineReducers } from 'redux';
+import { Provider } from 'react-redux';
+import rootReducer from './reducers';
+
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
+const store = createStore(rootReducer);
 
 ReactDOM.render(
   <React.StrictMode>
     <CookiesProvider>
-      <App />
+      <Provider store={store}>
+        <App />
+      </Provider>
     </CookiesProvider>
   </React.StrictMode>,
   document.getElementById('root')
