+++ client/src/pages/LoggedIn.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import 'bootstrap/dist/css/bootstrap.css';
+import "../styles/login.css";
+
+function Login() {
+  return (
+    <div className="loggedInTest">
+      <h1>Ay you're loggged IN</h1>
+      <h1>welcome 2 the club</h1>
+      <h2>hola</h2>
+    </div>
+  );
+}
+
+export default Login;
